+++ Changelog.md
@@ -100,3 +100,4 @@ _osl_panic for string "32KHz LPO Clock not running" does not produce panic in Ai
 #### v2.0.6
 - Add previous implementation of wowl disabler (both methods wlc_wowl_enable and wowCapablePlatform must be overridden)
 - Add support for DW1820A CN-096JNT & Fix freezing issues (thanks to Sniki)
+- Read property brcmfx-driver from provider's IOReg entry (can be injected by bootloder as well as brcmfx-country property)